pyproject.toml
[project]
name = "finance-assistant"
version = "0.1.0"
description = "Personal finance assistant backend — Costa Rica MVP"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.29.0",
    "alembic==1.13.3",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "redis==5.1.0",
    "httpx==0.27.2",
    "python-dateutil==2.9.0.post0",
    "anthropic==0.39.0",
    "aiogram==3.13.1",
    "rapidfuzz==3.10.1",
    "pyjwt==2.9.0",
]

# Optional production-only deps. Install via `uv sync --extra azure` when
# deploying with SECRET_STORE_BACKEND=azure_kv. Dev / tests never need these.
[project.optional-dependencies]
azure = [
    "azure-identity==1.19.0",
    "azure-keyvault-secrets==4.9.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
]

[tool.pytest.ini_options]
# Auto-apply @pytest.mark.asyncio isn't strictly required, but `auto` mode
# means we don't need to decorate every coroutine test. Safer for Phase 5b
# because extractor/dispatcher tests are nearly all async.
asyncio_mode = "auto"
testpaths = ["tests"]
